+++ projects/generic-components/src/lib/molecules/molecule-card/molecule-card.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { MoleculeCardComponent } from './molecule-card.component';
+import { AtomButtonModule } from '../../atoms/atom-button/atom-button.module';
+
+@NgModule({
+  declarations: [MoleculeCardComponent],
+  imports: [AtomButtonModule],
+  exports: [MoleculeCardComponent],
+})
+export class MoleculeCardModule {}