+++ projects/generic-components/src/lib/molecules/molecules.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { MoleculeCardModule } from './molecule-card/molecule-card.module';
+
+@NgModule({
+  imports: [MoleculeCardModule],
+  exports: [MoleculeCardModule],
+})
+export class MoleculesModule {}