+++ projects/generic-components/src/public-api.ts
@@ -8,3 +8,10 @@ export * from 'projects/generic-components/src/lib/atoms/atom-button/atom-button
 export * from 'projects/generic-components/src/lib/atoms/atom-button/atom-button.module'
 
 export * from 'projects/generic-components/src/lib/atoms/atoms.module'
+
+/* Molecules */
+
+export * from 'projects/generic-components/src/lib/molecules/molecule-card/molecule-card.component'
+export * from 'projects/generic-components/src/lib/molecules/molecule-card/molecule-card.module'
+
+export * from 'projects/generic-components/src/lib/molecules/molecules.module'