+++ src/app/components/my-users/my-users.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { MyUsersComponent } from './my-users.component';
-import { AtomsModule } from '@my-users/generic-components';
+import { AtomsModule, MoleculesModule } from '@my-users/generic-components';
 import { CommonModule } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [MyUsersComponent],
-  imports: [AtomsModule, CommonModule, BrowserModule, HttpClientModule],
+  imports: [AtomsModule, CommonModule, BrowserModule, HttpClientModule, MoleculesModule],
   exports: [MyUsersComponent],
 })
 export class MyUsersModule {}