+++ src/app/store/users/users.query.ts
@@ -1,5 +1,5 @@
 import { Query } from '@datorama/akita';
-import { UsersState, UsersStore } from '../users-store/users.store';
+import { UsersState, UsersStore } from './users.store';
 import { Injectable } from '@angular/core';
 
 @Injectable({ providedIn: 'root' })