+++ src/app/store/users/users.store.ts
@@ -1,16 +1,23 @@
 import { Store, StoreConfig } from '@datorama/akita';
-import { UserInterface } from '../../../../interfaces/user.interface';
-import {Injectable} from "@angular/core";
+import { UserInterface } from '../../interfaces/user.interface';
+import { Injectable } from '@angular/core';
 
 export interface UsersState {
   usersList: UserInterface[];
-  selectedUser: UserInterface | null;
+  selectedUser: UserInterface;
 }
 
 export function createInitialState(): UsersState {
   return {
     usersList: [],
-    selectedUser: null
+    selectedUser: {
+      id: '',
+      name: '',
+      username: '',
+      email: '',
+      phone: '',
+      website: '',
+    },
   };
 }