+++ projects/generic-components/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Features 
 
 - Add atom button component
+- Add molecule component
+- Set public-api file
 
 ## [0.0.1] - 2022-04-16